add all the test files to same folder
