+++ test/mockstorage.js
